requirements.txt
flask==2.2.0
flask-sqlalchemy==3.0.2
pandas==1.5.3
numpy==1.23.5
yfinance==0.2.31
scikit-learn==1.2.2
newsapi-python==0.2.7
python-dotenv==0.19.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
textblob==0.17.1
tensorflow==2.10.0
keras==2.10.0
h5py==3.8.0 